Extract kubelet node status into separate file
